Fix another int vs size_t warning.

+++ find_mode.cc
@@ -17,7 +17,7 @@ class FindMode : public EditorMode {
   bool SeekOnce(const shared_ptr<OpenBuffer> buffer, int c) {
     shared_ptr<LazyString> current_line = buffer->current_line()->contents;
     size_t line_length = current_line->size();
-    for (int i = buffer->current_position_col + 1; i < line_length; i++) {
+    for (size_t i = buffer->current_position_col + 1; i < line_length; i++) {
       if (current_line->get(i) == c) {
         buffer->current_position_col = i;
         return true;
